+++ tests/templates/hello.php
@@ -1,0 +1,1 @@
+Hello, <?=isset($hello)? $hello : 'World' ?>!