+++ tests/templates/layout3.php
@@ -1,0 +1,6 @@
+<html>
+<head>
+<title><?=$title?></title>
+</head>
+<body><?=$_content?></body>
+</html>