+++ tests/templates/setglobal.php
@@ -1,0 +1,2 @@
+<?php $this->setGlobal('title',$title)?>
+Hello, World!