+++ tests/templates/setlayout.php
@@ -1,0 +1,2 @@
+<?php $this->setLayout('layout.php') ?>
+Hello, World!