+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- `addHelper()` to add custom helpers
+
 ## [1.1.0] - 2017-03-25
 
 ### Added